fix(coding-memory): clippy lints + public_surface test update for Phase 3

+++ crates/coding-memory/src/code_domain_searcher.rs
@@ -58,8 +58,8 @@ impl DomainSearcher for CodeDomainSearcher {
                         id: m.id,
                         source: MemorySource::EpisodicMemory,
                         content: m.content,
-                        score: m.importance as f64,
-                        raw_score: m.importance as f64,
+                        score: m.importance,
+                        raw_score: m.importance,
                     });
                 }
             }
